+++ src/generate_all_from_schemas.sh
@@ -7,6 +7,7 @@ TEMPLATE_DIRECTORY=src/deployment_templates
 EXAMPLE_DIR=examples
 EXAMPLE_DATA_DIR=src/raw_example_data
 BLANK_DEPLOYMENT_DIR=$EXAMPLE_DIR/blank_deployment
+CSV_DEPLOYMENT_DIR=$EXAMPLE_DIR/csv_deployment
 MIN_BLANK_DEPLOYMENT_DIR=$EXAMPLE_DIR/minimum_required_blank_deployment
 
 if [[ -d schemas/  ]]; then
@@ -75,9 +76,14 @@ done
 echo "Generate deployment spreadsheet template CSV"
 python src/generate_deployment_spreadsheet_template.py > deployment_spreadsheet_template.csv
 
+
 echo "Generate deployment excel spreadsheet template"
 python src/convert_schemas_to_excel.py sample_deployment.xlsx
 python src/convert_schemas_to_excel.py $EXAMPLE_DIR/excel/blank_deployment.xlsx
+
+echo "Generate deployment spreadsheet example"
+mkdir $CSV_DEPLOYMENT_DIR
+python convert_deployment_to_csv.py examples/kvm_sdwan_install/ $CSV_DEPLOYMENT_DIR/sample_deployment.csv
 
 # Add specific files to default deployment
 add_to_default_deployment="common credentials upgrade vscs vsds vstats"